+++ packages/web3-eth-ens/test/integration/ens.test.ts
@@ -16,9 +16,8 @@ along with web3.js.  If not, see <http://www.gnu.org/licenses/>.
 */
 
 /* eslint-disable @typescript-eslint/no-unused-vars */
-import { getBlock } from 'web3-eth';
 import { Contract, PayableTxOptions } from 'web3-eth-contract';
-import { Address, Bytes, DEFAULT_RETURN_FORMAT } from 'web3-types';
+import { Address, Bytes } from 'web3-types';
 import { sha3, toChecksumAddress } from 'web3-utils';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { IpcProvider } from 'web3-providers-ipc';
@@ -76,7 +75,7 @@ describe('ens', () => {
 		const acc2 = await createTempAccount();
 		accountOne = acc2.address;
 
-		sendOptions = { from: defaultAccount, gas: '10000000' };
+		sendOptions = { from: defaultAccount, type: '0x1' };
 
 		const Registry = new Contract(ENSRegistryAbi, undefined, {
 			provider: getSystemTestProvider(),
@@ -121,15 +120,6 @@ describe('ens', () => {
 		else provider = new ENS.providers.HttpProvider(clientUrl);
 
 		ens = new ENS(registry.options.address, provider);
-
-		const block = await getBlock(ens, 'latest', false, DEFAULT_RETURN_FORMAT);
-		const gas = block.gasLimit.toString();
-
-		// Increase gas for contract calls
-		sendOptions = {
-			...sendOptions,
-			gas,
-		};
 	});
 
 	afterAll(async () => {