+++ packages/web3-utils/src/converters.ts
@@ -77,6 +77,8 @@ export const ethUnitMap = {
 	tether: BigInt('1000000000000000000000000000000'),
 };
 
+const PrecisionLossWarning = 'Warning: Using type `number` with values that are large or contain many decimals may cause loss of precision, it is recommended to use type `string` or `BigInt` when using conversion methods';
+
 export type EtherUnits = keyof typeof ethUnitMap;
 /**
  * Convert a value from bytes to Uint8Array
@@ -362,6 +364,10 @@ export const toHex = (
 		return returnType ? 'bigint' : numberToHex(value);
 	}
 
+	if(isUint8Array(value)) {
+		return returnType ? 'bytes' : bytesToHex(value);
+	}
+
 	if (typeof value === 'object' && !!value) {
 		return returnType ? 'string' : utf8ToHex(JSON.stringify(value));
 	}
@@ -411,7 +417,8 @@ export const toHex = (
  */
 export const toNumber = (value: Numbers): number | bigint => {
 	if (typeof value === 'number') {
-            if (value > 1e+20) {
+			if (value > 1e+20) {
+				console.warn(PrecisionLossWarning)
                 // JavaScript converts numbers >= 10^21 to scientific notation when coerced to strings,
                 // leading to potential parsing errors and incorrect representations.
                 // For instance, String(10000000000000000000000) yields '1e+22'.
@@ -551,17 +558,32 @@ export const toWei = (number: Numbers, unit: EtherUnits): string => {
 	if (!denomination) {
 		throw new InvalidUnitError(unit);
 	}
+	let parsedNumber = number;
+	if (typeof parsedNumber === 'number'){
+		if (parsedNumber  < 1e-15){
+			console.warn(PrecisionLossWarning)
+		}
+		if (parsedNumber > 1e+20) {
+			console.warn(PrecisionLossWarning)
 
+			parsedNumber =  BigInt(parsedNumber);
+		} else {
+			// in case there is a decimal point, we need to convert it to string
+			parsedNumber = parsedNumber.toLocaleString('fullwide', {useGrouping: false, maximumFractionDigits: 20})
+		}
+	}
+	
 	// if value is decimal e.g. 24.56 extract `integer` and `fraction` part
 	// to avoid `fraction` to be null use `concat` with empty string
 	const [integer, fraction] = String(
-		typeof number === 'string' && !isHexStrict(number) ? number : toNumber(number),
+		typeof parsedNumber === 'string' && !isHexStrict(parsedNumber) ? parsedNumber : toNumber(parsedNumber),
 	)
 		.split('.')
 		.concat('');
 
 	// join the value removing `.` from
 	// 24.56 -> 2456
+	
 	const value = BigInt(`${integer}${fraction}`);
 
 	// multiply value with denomination